+++ packages/firebase_auth/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.3.1
+
+* Change GMS dependency to 11.+
+
+## 0.3.0
+
+* **Breaking Change**: Method FirebaseUser getToken was renamed to getIdToken.
+
+## 0.2.5
+
+* Added support for linkWithCredential with Google credential
+
 ## 0.2.4
 
 * Added support for `signInWithCustomToken`