+++ packages/firebase_messaging/CHANGELOG.md
@@ -1,9 +1,18 @@
-## 0.0.6
+## 0.0.7
 
 In FirebaseMessagingPlugin.m:
 * moved logic from 'tokenRefreshNotification' to 'didRefreshRegistrationToken'
 * removed 'tokenRefreshNotification' as well as observer registration
 * removed 'connectToFcm' method and related calls
+* removed unnecessary FIRMessaging disconnect
+
+## 0.0.6
+
+* Change GMS dependency to 11.+
+
+## 0.0.5+2
+
+* Fixed README example for "click_action"
 
 ## 0.0.5+1